more refactoring of Utils.IO
